fix compiled error

+++ src/app/api/auth/article/delete/route.js
@@ -1,11 +1,15 @@
 import { NextResponse } from 'next/server';
+import { getServerSession } from 'next-auth';
 import dbConnect from '@/app/db/dbConnect';
 import Article from '@/app/db/models/Article';
+import { configAuth } from '../../[...nextauth]/route';
 
 export async function PUT(req) {
+  const session = await getServerSession(configAuth);
+
+  if (!session) return NextResponse.json({ error: 'Wrong Credentials' }, { status: 401 });
   dbConnect();
   const { id } = await req.json();
-  console.log(id);
   let article;
   try {
     article = await Article.findByIdAndDelete(id);
